(feat) sends image back to client!
Woot!! This sorta kinda works for the first time. Wahoo! It's also
ghetto as hell, but we walk before we fly.

This commit saves to the HTTP response object instead of the file
system. The tricksy bit is that the `save()` method from the image
library requires ownership of the writer, but we need to hand on
to the writer ownership to close the HTTP connection.

Yurume from the #rust-gamedev room suggested using the ByRefWriter
trait which essentially wraps the writer with a proxy writer. So,
I can have my writer and eat it too.